chore: Make yarn install modules in node=10
Yarn, as of yet, does not fully support node=10.
But in order for us to test in that environment we should be able to install
modules in node=10.

So redefine installation instructions for the yarn in CI environment.

Also make git ignore yarn logs as well as regenerate the changelog